Calculate of Goldbach and Fact

+++ project1-1/src/productionLine/Calculator.hpp
@@ -2,6 +2,7 @@
 #include "../http/HttpPendingRequest.hpp"
 #include <vector>
 
+
 class Calculator {
  protected:
         // Number to calculate
@@ -20,7 +21,7 @@ class Calculator {
         virtual ~Calculator() {}
 
         // Calculate the number
-        virtual std::vector<int> calculate(int number) = 0;
+        virtual std::vector<int64_t> calculate(int number) = 0;
 
         /// Get calcIndex
         int64_t getCalcIndex() const {
