+++ project1-1/src/productionLine/GoldCalculator.hpp
@@ -1,7 +1,9 @@
 // Copyright 2024 Aaron Santana, Isaias Alfaro
 #pragma once
 
+#include <vector>
 #include "Calculator.hpp"
+#include "../model/GoldCal.hpp"
 
 class GoldCalculator : public Calculator {
  public:
@@ -9,5 +11,5 @@ class GoldCalculator : public Calculator {
       Calculator(calcIndex, pendingRequest) {}
   ~GoldCalculator();
 
-  std::vector<int> calculate(int number) override;
+  std::vector<int64_t> calculate(int number) override;
 };
